requirements.txt
discord.py==7.46.2
